% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaBinary_generics.R
\name{AIC.hpaBinary}
\alias{AIC.hpaBinary}
\title{Calculates AIC for "hpaBinary" object}
\usage{
\method{AIC}{hpaBinary}(object, ..., k = 2)
}
\arguments{
\item{object}{Object of class "hpaBinary"}

\item{...}{further arguments (currently ignored)}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is 
the classical AIC.}
}
\description{
This function calculates AIC for "hpaBinary" object
}
