% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaML_generics.R
\name{AIC.hpaML}
\alias{AIC.hpaML}
\title{Calculates AIC for "hpaML" object}
\usage{
\method{AIC}{hpaML}(object, ..., k = 2)
}
\arguments{
\item{object}{Object of class "hpaML"}

\item{...}{further arguments (currently ignored)}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is 
the classical AIC.}
}
\description{
This function calculates AIC for "hpaML" object
}
