% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AIC.hpaSelection}
\alias{AIC.hpaSelection}
\title{Calculates AIC for "hpaSelection" object}
\usage{
\method{AIC}{hpaSelection}(model)
}
\description{
This function calculates AIC for "hpaSelection" object
}
