% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AIC_hpaBinary}
\alias{AIC_hpaBinary}
\title{Calculates AIC for "hpaBinary" object}
\usage{
AIC_hpaBinary(object, k = 2)
}
\arguments{
\item{object}{Object of class "hpaBinary"}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is 
the classical AIC.}
}
\description{
This function calculates AIC for "hpaBinary" object
}
