% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AIC_hpaML}
\alias{AIC_hpaML}
\title{Calculates AIC for "hpaML" object}
\usage{
AIC_hpaML(object, k = 2)
}
\arguments{
\item{object}{Object of class "hpaML"}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is 
the classical AIC.}
}
\description{
This function calculates AIC for "hpaML" object
}
