% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AIC_hpaSelection}
\alias{AIC_hpaSelection}
\title{Calculates AIC for "hpaSelection" object}
\usage{
AIC_hpaSelection(object, k = 2)
}
\arguments{
\item{object}{Object of class "hpaSelection"}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is 
the classical AIC.}
}
\description{
This function calculates AIC for "hpaSelection" object
}
