% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaBinary_generics.R
\name{coef.hpaBinary}
\alias{coef.hpaBinary}
\title{Extract coefficients from hpaBinary object}
\usage{
\method{coef}{hpaBinary}(object, ...)
}
\arguments{
\item{object}{Object of class "hpaBinary"}

\item{...}{further arguments (currently ignored)}
}
\description{
Extract coefficients from hpaBinary object
}
