% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaML_generics.R
\name{coef.hpaML}
\alias{coef.hpaML}
\title{Extract coefficients from hpaML object}
\usage{
\method{coef}{hpaML}(object, ...)
}
\arguments{
\item{object}{Object of class "hpaML"}

\item{...}{further arguments (currently ignored)}
}
\description{
Extract coefficients from hpaML object
}
