% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLik.hpaBinary}
\alias{logLik.hpaBinary}
\title{Calculates log-likelihood for "hpaBinary" object}
\usage{
logLik.hpaBinary(model)
}
\description{
This function calculates log-likelihood for "hpaBinary" object
}
