% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaML_generics.R
\name{logLik.hpaML}
\alias{logLik.hpaML}
\title{Calculates log-likelihood for "hpaML" object}
\usage{
\method{logLik}{hpaML}(object, ...)
}
\arguments{
\item{object}{Object of class "hpaML"}

\item{...}{further arguments (currently ignored)}
}
\description{
This function calculates log-likelihood for "hpaML" object
}
