% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLik.hpaSelection}
\alias{logLik.hpaSelection}
\title{Calculates log-likelihood for "hpaSelection" object}
\usage{
logLik.hpaSelection(model)
}
\description{
This function calculates log-likelihood for "hpaSelection" object
}
