% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLik_hpaBinary}
\alias{logLik_hpaBinary}
\title{Calculates log-likelihood for "hpaBinary" object}
\usage{
logLik_hpaBinary(object)
}
\arguments{
\item{object}{Object of class "hpaBinary"}
}
\description{
This function calculates log-likelihood for "hpaBinary" object
}
