% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLik_hpaML}
\alias{logLik_hpaML}
\title{Calculates log-likelihood for "hpaML" object}
\usage{
logLik_hpaML(object)
}
\arguments{
\item{object}{Object of class "hpaML"}
}
\description{
This function calculates log-likelihood for "hpaML" object
}
