% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLik_hpaSelection}
\alias{logLik_hpaSelection}
\title{Calculates log-likelihood for "hpaSelection" object}
\usage{
logLik_hpaSelection(object)
}
\arguments{
\item{object}{Object of class "hpaSelection"}
}
\description{
This function calculates log-likelihood for 
"hpaSelection" object
}
