% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lseq.R
\name{lseq}
\alias{lseq}
\title{\code{lseq} output a sequence of logarithmically spaced numbers.}
\usage{
lseq(from, to, length.out = 10)
}
\arguments{
\item{from}{first number in the sequence.}

\item{to}{last number in the sequence.}

\item{length.out}{length of the sequence, a positive integer > 1. Default is 10.}
}
\value{
A numeric vector.
}
\description{
\code{lseq} works like \code{seq} in that it returns a vector of numbers with \code{from} as the first element, \code{to} as the last elelemnt, and \code{length.out} as the length of the vector, but the values are logarithmically spaced.
}
\examples{
lseq(0.1, 10, length.out = 20)   # 20 log-spaced values between 0.1 and 10
# plot a log-spaced grid
lim <- c(0.1, 10)
plot(1, 1, xlim=lim, ylim=lim, type="n", axes=FALSE, xlab="", ylab="")
x <- lseq(0.1, 10, length.out = 10)
abline(v = x, col="#3366FF66")
abline(h = x, col="#3366FF66")
}
\seealso{
\code{\link{seq}}
}
