% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mecdf}
\alias{mecdf}
\title{Calculates multivariate empirical cumulative distribution function}
\usage{
mecdf(x)
}
\arguments{
\item{x}{numeric matrix which rows are observations}
}
\description{
This function calculates multivariate 
empirical cumulative distribution function
at each point of the sample
}
