% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaBinary_generics.R
\name{plot.hpaBinary}
\alias{plot.hpaBinary}
\title{Plot the approximated density of hpaBinary random errors}
\usage{
\method{plot}{hpaBinary}(x, y = NULL, ...)
}
\arguments{
\item{x}{Object of class "hpaBinary"}

\item{y}{this parameter currently ignored}

\item{...}{further arguments to be passed to the \code{\link[base]{plot}}
function.}
}
\description{
Plot the approximated density of hpaBinary random errors
}
