% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{plot.hpaSelection}
\alias{plot.hpaSelection}
\title{Plot hpaSelection random errors approximated density}
\usage{
plot.hpaSelection(model, is_outcome = TRUE)
}
\arguments{
\item{model}{Object of class "hpaSelection"}

\item{is_outcome}{logical; if TRUE then function plots the graph for outcome equation random errors. 
Otherwise plot for selection equation random errors will be plotted.}
}
\description{
Plot hpaSelection random errors approximated density
}
