% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{plot_hpaBinary}
\alias{plot_hpaBinary}
\title{Plot hpaBinary random errors approximated density}
\usage{
plot_hpaBinary(x)
}
\arguments{
\item{x}{Object of class "hpaBinary"}
}
\description{
Plot hpaBinary random errors approximated density
}
