% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaML_generics.R
\name{predict.hpaML}
\alias{predict.hpaML}
\title{Predict method for hpaML}
\usage{
\method{predict}{hpaML}(object, ..., newdata = matrix(c(0)))
}
\arguments{
\item{object}{Object of class "hpaML"}

\item{...}{further arguments (currently ignored)}

\item{newdata}{An optional data frame (for \link[hpa]{hpaBinary} and
\link[hpa]{hpaSelection}) or numeric matrix (for \link[hpa]{hpaML})
in which to look for variables with which to predict. If omitted,
the original data frame (matrix) used.}
}
\value{
This function returns predictions based on 
\code{\link[hpa]{hpaML}} estimation results.
}
\description{
Predict method for hpaML
}
