% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_hpaML}
\alias{predict_hpaML}
\title{Predict method for hpaML}
\usage{
predict_hpaML(object, newdata = matrix(1, 1))
}
\arguments{
\item{object}{Object of class "hpaML"}

\item{newdata}{An optional data frame (for \link[hpa]{hpaBinary} and
\link[hpa]{hpaSelection}) or numeric matrix (for \link[hpa]{hpaML})
in which to look for variables with which to predict. If omitted,
the original data frame (matrix) used.}
}
\value{
This function returns predictions based 
on \code{\link[hpa]{hpaML}} estimation results.
}
\description{
Predict method for hpaML
}
