% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaBinary_generics.R
\name{print.hpaBinary}
\alias{print.hpaBinary}
\title{Print method for "hpaBinary" object}
\usage{
\method{print}{hpaBinary}(x, ...)
}
\arguments{
\item{x}{Object of class "hpaBinary"}

\item{...}{further arguments (currently ignored)}
}
\description{
Print method for "hpaBinary" object
}
