% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaML_generics.R
\name{print.hpaML}
\alias{print.hpaML}
\title{Print method for "hpaML" object}
\usage{
\method{print}{hpaML}(x, ...)
}
\arguments{
\item{x}{Object of class "hpaML"}

\item{...}{further arguments (currently ignored)}
}
\description{
Print method for "hpaML" object
}
