% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaSelection_generics.R
\name{print.hpaSelection}
\alias{print.hpaSelection}
\title{Print method for "hpaSelection" object}
\usage{
\method{print}{hpaSelection}(x, ...)
}
\arguments{
\item{x}{Object of class "hpaSelection"}

\item{...}{further arguments (currently ignored)}
}
\description{
Print method for "hpaSelection" object
}
