% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaBinary_generics.R
\name{print.summary.hpaBinary}
\alias{print.summary.hpaBinary}
\title{Summary for "hpaBinary" object}
\usage{
\method{print}{summary.hpaBinary}(x, ...)
}
\arguments{
\item{x}{Object of class "hpaBinary"}

\item{...}{further arguments (currently ignored)}
}
\description{
Summary for "hpaBinary" object
}
