% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print.summary.hpaML}
\alias{print.summary.hpaML}
\title{Summary for hpaML output}
\usage{
\method{print}{summary.hpaML}(model)
}
\arguments{
\item{model}{Object of class "hpaML"}
}
\description{
Summary for hpaML output
}
