% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaSelection_generics.R
\name{print.summary.hpaSelection}
\alias{print.summary.hpaSelection}
\title{Summary for "hpaSelection" object}
\usage{
\method{print}{summary.hpaSelection}(x, ...)
}
\arguments{
\item{x}{Object of class "hpaSelection"}

\item{...}{further arguments (currently ignored)}
}
\description{
Summary for "hpaSelection" object
}
