% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print_summary_hpaBinary}
\alias{print_summary_hpaBinary}
\title{Summary for hpaBinary output}
\usage{
print_summary_hpaBinary(x)
}
\arguments{
\item{x}{Object of class "hpaML"}
}
\description{
Summary for hpaBinary output
}
