% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print_summary_hpaML}
\alias{print_summary_hpaML}
\title{Summary for hpaML output}
\usage{
print_summary_hpaML(x)
}
\arguments{
\item{x}{Object of class "hpaML"}
}
\description{
Summary for hpaML output
}
