% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printPolynomial}
\alias{printPolynomial}
\title{Print polynomial given it's degrees and coefficients}
\usage{
printPolynomial(pol_degrees, pol_coefficients)
}
\arguments{
\item{pol_degrees}{non-negative integer vector of polynomial degrees.}

\item{pol_coefficients}{numeric vector of polynomial coefficients.}
}
\value{
This function returns the string which contains polynomial symbolic representation.
}
\description{
This function prints polynomial given it's degrees and coefficients.
}
\details{
Function automatically removes polynomial elements which coefficient are zero
and variables which power is zero. Output may contain long coefficients representation as they
are not rounded.
}
\examples{
##Let's represent polynomial 0.3+0.5x2-x2^2+2x1+1.5x1x2+x1x2^2

pol_degrees <- c(1, 2)
pol_coefficients <- c(0.3, 0.5, -1, 2, 1.5, 1)

printPolynomial(pol_degrees, pol_coefficients)
}
