% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{summary.hpaBinary}
\alias{summary.hpaBinary}
\title{Summarizing hpaBinary Fits}
\usage{
\method{summary}{hpaBinary}(model)
}
\arguments{
\item{model}{Object of class "hpaBinary"}
}
\value{
This function returns the same list as \code{\link[hpa]{hpaBinary}} function changing it's class to "summary.hpaBinary".
}
\description{
Summarizing hpaBinary Fits
}
