% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaML_generics.R
\name{summary.hpaML}
\alias{summary.hpaML}
\title{Summarizing hpaML Fits}
\usage{
\method{summary}{hpaML}(object, ...)
}
\arguments{
\item{object}{Object of class "hpaML"}

\item{...}{further arguments (currently ignored)}
}
\value{
This function returns the same list as \code{\link[hpa]{hpaML}} 
function changing its class to "summary.hpaML".
}
\description{
Summarizing hpaML Fits
}
