% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{summary.hpaSelection}
\alias{summary.hpaSelection}
\title{Summarizing hpaSelection Fits}
\usage{
\method{summary}{hpaSelection}(model)
}
\arguments{
\item{model}{Object of class "hpaSelection"}
}
\value{
This function returns the same list as \code{\link[hpa]{hpaSelection}} function changing it's class to "summary.hpaSelection".
}
\description{
This function summarizing hpaSelection Fits
}
