% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{summary_hpaBinary}
\alias{summary_hpaBinary}
\title{Summarizing hpaBinary Fits}
\usage{
summary_hpaBinary(object)
}
\arguments{
\item{object}{Object of class "hpaBinary"}
}
\value{
This function returns the same list as 
\code{\link[hpa]{hpaBinary}} function changing 
its class to "summary.hpaBinary".
}
\description{
Summarizing hpaBinary Fits
}
