% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{summary_hpaML}
\alias{summary_hpaML}
\title{Summarizing hpaML Fits}
\usage{
summary_hpaML(object)
}
\arguments{
\item{object}{Object of class "hpaML"}
}
\value{
This function returns the same 
list as \code{\link[hpa]{hpaML}} function changing 
its class to "summary.hpaML".
}
\description{
Summarizing hpaML Fits
}
