% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{summary_hpaSelection}
\alias{summary_hpaSelection}
\title{Summarizing hpaSelection Fits}
\usage{
summary_hpaSelection(object)
}
\arguments{
\item{object}{Object of class "hpaSelection".}
}
\value{
This function returns the same list as 
\code{\link[hpa]{hpaSelection}} function changing its class 
to "summary.hpaSelection".
}
\description{
This function summarizing hpaSelection Fits.
}
