% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaBinary_generics.R
\name{vcov.hpaBinary}
\alias{vcov.hpaBinary}
\title{Extract covariance matrix from hpaBinary object}
\usage{
\method{vcov}{hpaBinary}(object, ...)
}
\arguments{
\item{object}{Object of class "hpaBinary"}

\item{...}{further arguments (currently ignored)}
}
\description{
Extract covariance matrix from hpaBinary object
}
