% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaML_generics.R
\name{vcov.hpaML}
\alias{vcov.hpaML}
\title{Extract covariance matrix from hpaML object}
\usage{
\method{vcov}{hpaML}(object, ...)
}
\arguments{
\item{object}{Object of class "hpaML"}

\item{...}{further arguments (currently ignored)}
}
\description{
Extract covariance matrix from hpaML object
}
