% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaSelection_generics.R
\name{vcov.hpaSelection}
\alias{vcov.hpaSelection}
\title{Extract covariance matrix from hpaSelection object}
\usage{
\method{vcov}{hpaSelection}(object, ...)
}
\arguments{
\item{object}{Object of class "hpaSelection"}

\item{...}{further arguments (currently ignored)}
}
\description{
Extract covariance matrix from hpaSelection object
}
