% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaSelection_generics.R
\name{AIC.hpaSelection}
\alias{AIC.hpaSelection}
\title{Calculates AIC for "hpaSelection" object}
\usage{
\method{AIC}{hpaSelection}(object, ..., k = 2)
}
\arguments{
\item{object}{Object of class "hpaSelection"}

\item{...}{further arguments (currently ignored)}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is 
the classical AIC.}
}
\description{
This function calculates AIC for "hpaSelection" object
}
