% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaSelection_generics.R
\name{coef.hpaSelection}
\alias{coef.hpaSelection}
\title{Extract coefficients from hpaSelection object}
\usage{
\method{coef}{hpaSelection}(object, ..., type = "all")
}
\arguments{
\item{object}{Object of class "hpaSelection"}

\item{...}{further arguments (currently ignored)}

\item{type}{character; if "all" (default) then all estimated parameters
values will be returned. If "selection" then selection equation coefficients
estimates will be provided. If "outcome" then outcome equation coefficients
estimates will be returned.}
}
\description{
Extract coefficients from hpaSelection object
}
