% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnorm_parallel}
\alias{dnorm_parallel}
\title{Calculate normal pdf in parallel}
\usage{
dnorm_parallel(x, mean = 0, sd = 1, is_parallel = FALSE)
}
\arguments{
\item{x}{numeric vector of quantiles.}

\item{mean}{double value.}

\item{sd}{double positive value.}

\item{is_parallel}{if \code{TRUE} then multiple cores will be
used for some calculations. It usually provides speed advantage for
large enough samples (about more than 1000 observations).}
}
\description{
Calculate in parallel for each value from vector \code{x} 
density function of normal distribution with 
mean equal to \code{mean} and standard deviation equal to \code{sd}.
}
\examples{
## Consider normal distribution with mean 3 and standard deviation 5.
## Calculate its density function at points 2 and 3.

# Create vector of points
my_points <- c(2, 3)

# Calculate pdf at these points 
# (set is_parallel = TRUE in order 
# to turn on parallel computations)
dnorm_parallel(my_points, 3, 5, 
               is_parallel = FALSE)
}
