% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaBinary_generics.R
\name{logLik.hpaBinary}
\alias{logLik.hpaBinary}
\title{Calculates log-likelihood for "hpaBinary" object}
\usage{
\method{logLik}{hpaBinary}(object, ...)
}
\arguments{
\item{object}{Object of class "hpaBinary"}

\item{...}{further arguments (currently ignored)}
}
\description{
This function calculates log-likelihood for "hpaBinary" object
}
