% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaSelection_generics.R
\name{logLik.hpaSelection}
\alias{logLik.hpaSelection}
\title{Calculates log-likelihood for "hpaSelection" object}
\usage{
\method{logLik}{hpaSelection}(object, ...)
}
\arguments{
\item{object}{Object of class "hpaSelection"}

\item{...}{further arguments (currently ignored)}
}
\description{
This function calculates log-likelihood for "hpaSelection" object
}
