% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaBinary_generics.R
\name{plot.hpaBinary}
\alias{plot.hpaBinary}
\title{Plot hpaBinary random errors approximated density}
\usage{
\method{plot}{hpaBinary}(x, y = NULL, ...)
}
\arguments{
\item{x}{Object of class "hpaBinary"}

\item{y}{this parameter currently ignored}

\item{...}{further arguments (currently ignored)}
}
\description{
Plot hpaBinary random errors approximated density
}
