% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{plot.hpaBinary}
\alias{plot.hpaBinary}
\title{Plot hpaBinary random errors approximated density}
\usage{
plot.hpaBinary(model)
}
\arguments{
\item{model}{Object of class "hpaBinary"}
}
\description{
Plot hpaBinary random errors approximated density
}
