% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaSelection_generics.R
\name{plot.hpaSelection}
\alias{plot.hpaSelection}
\title{Plot hpaSelection random errors approximated density}
\usage{
\method{plot}{hpaSelection}(x, y = NULL, ..., type = "outcome")
}
\arguments{
\item{x}{Object of class "hpaSelection"}

\item{y}{this parameter currently ignored}

\item{...}{further arguments to be passed to \code{\link[base]{plot}}
function.}

\item{type}{character; if "outcome" then function plots the graph for 
outcome equation random errors, if "selection" then plot for selection 
equation random errors will be generated.}
}
\description{
Plot hpaSelection random errors approximated density
}
