% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print.summary.hpaBinary}
\alias{print.summary.hpaBinary}
\title{Summary for hpaBinary output}
\usage{
\method{print}{summary.hpaBinary}(model)
}
\arguments{
\item{model}{Object of class "hpaML"}
}
\description{
Summary for hpaBinary output
}
