% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaML_generics.R
\name{print.summary.hpaML}
\alias{print.summary.hpaML}
\title{Summary for hpaML output}
\usage{
\method{print}{summary.hpaML}(x, ...)
}
\arguments{
\item{x}{Object of class "hpaML"}

\item{...}{further arguments (currently ignored)}
}
\description{
Summary for hpaML output
}
