% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print.summary.hpaSelection}
\alias{print.summary.hpaSelection}
\title{Summary for hpaSelection output}
\usage{
\method{print}{summary.hpaSelection}(model)
}
\arguments{
\item{model}{Object of class "hpaSelection"}
}
\description{
Summary for hpaSelection output
}
