% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaBinary_generics.R
\name{summary.hpaBinary}
\alias{summary.hpaBinary}
\title{Summarizing hpaBinary Fits}
\usage{
\method{summary}{hpaBinary}(object, ...)
}
\arguments{
\item{object}{Object of class "hpaBinary"}

\item{...}{further arguments (currently ignored)}
}
\value{
This function returns the same list as \code{\link[hpa]{hpaBinary}} 
function changing its class to "summary.hpaBinary".
}
\description{
Summarizing hpaBinary Fits
}
