% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpaSelection_generics.R
\name{summary.hpaSelection}
\alias{summary.hpaSelection}
\title{Summarizing hpaSelection Fits}
\usage{
\method{summary}{hpaSelection}(object, ...)
}
\arguments{
\item{object}{Object of class "hpaSelection"}

\item{...}{further arguments (currently ignored)}
}
\value{
This function returns the same list as \code{\link[hpa]{hpaSelection}} 
function changing its class to "summary.hpaSelection".
}
\description{
This function summarizing hpaSelection Fits
}
